pyproject.toml
[tool.poetry]
name = "aia-read-svc"
version = "0.1.0"
description = ""
authors = ["Edgar Rodrigo Sanchez De Gracia <edgr.sanchez@mercadolibre.cl>"]
readme = "README.md"
packages = [{include = "aia_read_svc"}]

[tool.poetry.dependencies]
python = ">=3.11.4,<3.12"
selenium = "^4.14.0"
numpy = "^1.24.0"
pywhatkit = "^5.4"
whatsapp-web = "^0.0.1"
confluent-kafka = "2.3.0"
pymongo = "4.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-dotenv = "^0.5.2"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
daemon = 'aia_read_svc.daemon:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
